+++ packages/react/src/index.js
@@ -9,6 +9,7 @@ export { Breadcrumb } from "./components/breadcrumb/Breadcrumb.jsx";
 export { Notification } from "./components/notification/Notification.jsx";
 export { Snackbar } from "./components/snackbar/Snackbar.jsx";
 export { StatusBar } from "./components/status-bar/StatusBar.jsx";
+export { Stat } from "./components/stat/Stat.jsx";
 export { CircularProgress } from "./components/circular-progress/CircularProgress.jsx";
 export { StepTracker } from "./components/step-tracker/StepTracker.jsx";
 export { Bleed } from "./components/bleed/Bleed.jsx";